Moved the text_type() method from scanner into T::Tokens so that everything can access it.
Added the HTML tag to control what kind of elements get created to represent
text chunks: text or html.

Added the Template::TT3::HTML module which monkey-patches various other modules
with html() methods.  This allows us to automaticaly HTML-encode values.

Added the 'raw' and 'block' commands.  Fixed up the grammar so that we can
have a 'block' command that doesn't clash with the basic 'block' element.

Refactored the Scope and Context so callbacks to the scanner and/or output
stream can be cleaner and better co-ordinated.